webview: fix type of upload_data_size for microhttpd
The upload_data_size is not a long unsigned * but a size_t *. These are
the same on x86_64, but not on i686.

+++ src/libs/webview/request_dispatcher.cpp
@@ -170,7 +170,7 @@ WebRequestDispatcher::process_request_cb(void *                 callback_data,
                                          const char *           method,
                                          const char *           version,
                                          const char *           upload_data,
-                                         long unsigned int *    upload_data_size,
+                                         size_t *               upload_data_size,
                                          void **                session_data)
 {
 	WebRequestDispatcher *rd = static_cast<WebRequestDispatcher *>(callback_data);
